requirement.txt
streamlit
PyPDF2
sentence-transformers
chromadb
litellm
langchain
langchain-community
python-dotenv
huggingface_hub
